update 401 error logic

+++ frontend/src/config/interceptors/handleError.ts
@@ -1,10 +1,11 @@
 import { AxiosError } from 'axios'
 import { userRepository } from 'config/userRepository'
+import { AppNavigationRoutes } from 'constants/paths'
 import { toastMessage } from 'utils/toastMessage'
 
 export const handleError = (error: AxiosError): void => {
   if (error.response) {
-    if (error.response.status === 401) {
+    if (error.response.status === 401 && window.location.pathname !== AppNavigationRoutes.SignIn) {
       userRepository.removeUser()
       window.location.href = 'login'
     }
